todo/requirements.txt
aniso8601==9.0.1
atomicwrites==1.4.1
attrs==22.1.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
PyJWT==1.7.1
pyparsing==3.0.9
pytest==7.1.2
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.40
tomli==2.0.1
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
